Änderungen aus besprechung mit Robin

+++ DailyData/app/src/main/java/com/pseandroid2/dailydata/repository/viewModelAPI/communicationClasses/ProjectPreview.kt
@@ -23,11 +23,11 @@ package com.pseandroid2.dailydata.repository.viewModelAPI.communicationClasses
 import com.pseandroid2.dailydata.model.database.entities.ProjectData
 
 class ProjectPreview(
-    override val id: Long,
+    override val id: Int,
     val name: String,
     val image: String
 ): Identifiable {
-    constructor(projectData: ProjectData) : this(projectData.id.toLong(), projectData.name, projectData.wallpaper)
+    constructor(projectData: ProjectData) : this(projectData.id, projectData.name, projectData.wallpaper)
     //@throws IllegalOperationException
     override suspend fun delete() {
         TODO("Not yet implemented")
